chore(ast): move ops to seperate file
